second clean up
