+++ apps/catalog/catalog/src/app/dashboard/title/details/details.module.ts
@@ -5,19 +5,19 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { TitleDetailsComponent } from './details.component';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { MovieSummaryMainModule } from '@blockframes/movie/form/summary/main/main.module';
-import { MovieSummaryFestivalPrizesModule } from '@blockframes/movie/form/summary/festival-prizes/festival-prizes.module';
-import { MovieSummarySalesCastModule } from '@blockframes/movie/form/summary/sales-cast/sales-cast.module';
-import { MovieSummaryCountryModule } from '@blockframes/movie/form/summary/country/country.module';
-import { MovieSummaryInformationModule } from '@blockframes/movie/form/summary/information/information.module';
-import { MovieSummaryStoryModule } from '@blockframes/movie/form/summary/story/story.module';
-import { MovieSummaryCreditModule } from '@blockframes/movie/form/summary/credit/credit.module';
-import { MovieSummaryImageModule } from '@blockframes/movie/form/summary/image/image.module';
-import { MovieSummaryBudgetModule } from '@blockframes/movie/form/summary/budget/budget.module';
-import { MovieSummaryTechnicalInformationModule } from '@blockframes/movie/form/summary/technical-information/technical-information.module';
-import { MovieSummaryEvaluationModule } from '@blockframes/movie/form/summary/evaluation/evaluation.module';
-import { MovieSummaryFileModule } from '@blockframes/movie/form/summary/file/file.module';
+
+import { MovieSummaryMainModule } from '@blockframes/movie/dashboard/components/summary/main/main.module';
+import { MovieSummaryFestivalPrizesModule } from '@blockframes/movie/dashboard/components/summary/festival-prizes/festival-prizes.module';
+import { MovieSummarySalesCastModule } from '@blockframes/movie/dashboard/components/summary/sales-cast/sales-cast.module';
+import { MovieSummaryCountryModule } from '@blockframes/movie/dashboard/components/summary/country/country.module';
+import { MovieSummaryInformationModule } from '@blockframes/movie/dashboard/components/summary/information/information.module';
+import { MovieSummaryStoryModule } from '@blockframes/movie/dashboard/components/summary/story/story.module';
+import { MovieSummaryCreditModule } from '@blockframes/movie/dashboard/components/summary/credit/credit.module';
+import { MovieSummaryImageModule } from '@blockframes/movie/dashboard/components/summary/image/image.module';
+import { MovieSummaryBudgetModule } from '@blockframes/movie/dashboard/components/summary/budget/budget.module';
+import { MovieSummaryTechnicalInformationModule } from '@blockframes/movie/dashboard/components//summary/technical-information/technical-information.module';
+import { MovieSummaryEvaluationModule } from '@blockframes/movie/dashboard/components/summary/evaluation/evaluation.module';
+import { MovieSummaryFileModule } from '@blockframes/movie/dashboard/components/summary/file/file.module';
 
 
 
@@ -26,9 +26,11 @@ import { MovieSummaryFileModule } from '@blockframes/movie/form/summary/file/fil
   imports: [
     CommonModule,
     FlexLayoutModule,
+
+    // Material
     MatCardModule,
-    MatIconModule,
     MatButtonModule,
+
     // Summary components
     MovieSummaryMainModule,
     MovieSummaryFestivalPrizesModule,
