+++ libs/movie/src/lib/movie/dashboard/components/summary/budget/budget.component.ts
@@ -1,8 +1,8 @@
 import { Component, ChangeDetectionStrategy, Input, ChangeDetectorRef, OnInit } from '@angular/core';
 import { MovieReview } from '@blockframes/movie/+state/movie.model';
 import { FormList } from '@blockframes/utils/form/forms/list.form';
-import { MovieReviewForm } from '../../movie.form';
-import { MovieForm } from '../../movie.form';
+import { MovieReviewForm } from '../../../../form/movie.form';
+import { MovieForm } from '../../../../form/movie.form';
 
 @Component({
   selector: '[movie] [review] movie-summary-budget',
@@ -15,7 +15,7 @@ export class MovieSummaryBudgetComponent implements OnInit {
   @Input() review: FormList<MovieReview>;
   @Input() link: string;
 
-  constructor(private cdr: ChangeDetectorRef) {}
+  constructor(private cdr: ChangeDetectorRef) { }
 
   ngOnInit() {
     this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
