+++ libs/movie/src/lib/movie/dashboard/components/summary/country/country.component.ts
@@ -1,0 +1,13 @@
+import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
+import { MovieForm } from '../../../../form/movie.form';
+
+@Component({
+  selector: '[movie] movie-summary-country',
+  templateUrl: './country.component.html',
+  styleUrls: ['./country.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class MovieSummaryCountryComponent {
+  @Input() movie: MovieForm;
+  @Input() link: string;
+}
