+++ libs/movie/src/lib/movie/dashboard/components/summary/credit/credit.component.scss
