+++ libs/movie/src/lib/movie/dashboard/components/summary/evaluation/evaluation.component.scss
