+++ libs/movie/src/lib/movie/dashboard/components/summary/festival-prizes/festival-prizes.component.html
