+++ libs/movie/src/lib/movie/dashboard/components/summary/file/file.component.scss
