+++ libs/movie/src/lib/movie/dashboard/components/summary/image/image.component.html
