+++ libs/movie/src/lib/movie/dashboard/components/summary/information/information.component.scss
