+++ libs/movie/src/lib/movie/dashboard/components/summary/main/main.module.ts
