+++ libs/movie/src/lib/movie/dashboard/components/summary/sales-cast/sales-cast.component.scss
