+++ libs/movie/src/lib/movie/dashboard/components/summary/story/story.component.scss
