+++ libs/movie/src/lib/movie/dashboard/components/summary/story/story.component.ts
@@ -1,6 +1,9 @@
+// Angular
 import { ENTER, COMMA } from '@angular/cdk/keycodes';
-import { Component, ChangeDetectionStrategy, Input, ChangeDetectorRef, OnInit } from '@angular/core';
-import { MovieForm } from '../../movie.form';
+import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
+
+// Form
+import { MovieForm } from '../../../../form/movie.form';
 
 @Component({
   selector: '[movie] movie-summary-story',
@@ -8,16 +11,10 @@ import { MovieForm } from '../../movie.form';
   styleUrls: ['./story.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class MovieSummaryStoryComponent implements OnInit {
+export class MovieSummaryStoryComponent {
 
   public separatorKeysCodes = [ENTER, COMMA]
 
   @Input() movie: MovieForm;
   @Input() link: string;
-
-  constructor(private cdr: ChangeDetectorRef) {}
-
-  ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
-  }
 }
