+++ libs/movie/src/lib/movie/dashboard/components/summary/technical-information/technical-information.component.ts
@@ -1,11 +1,17 @@
+// Angular
 import {
   Component,
   ChangeDetectionStrategy,
   Input,
   ChangeDetectorRef,
-  OnInit
+  OnInit,
+  OnDestroy
 } from '@angular/core';
-import { MovieForm } from '../../movie.form';
+
+import { MovieForm } from '../../../../form/movie.form';
+
+// RxJs
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: '[movie] movie-summary-technical-information',
@@ -13,17 +19,23 @@ import { MovieForm } from '../../movie.form';
   styleUrls: ['./technical-information.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class MovieSummaryTechnicalInformationComponent implements OnInit {
+export class MovieSummaryTechnicalInformationComponent implements OnInit, OnDestroy {
   @Input() movie: MovieForm;
   @Input() link: string;
+
+  private sub: Subscription;
 
   constructor(private cdr: ChangeDetectorRef) { }
 
   ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
+    this.sub = this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
   }
 
   get hasKeys() {
     return Object.keys(this.movie.controls).length;
+  }
+
+  ngOnDestroy() {
+    if (this.sub) this.sub.unsubscribe();
   }
 }
