+++ libs/movie/src/lib/movie/dashboard/components/summary/technical-information/technical-information.module.ts
