--- libs/movie/src/lib/movie/form/summary-deprecated/country/country.component.ts
@@ -1,19 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input, ChangeDetectorRef, OnInit } from '@angular/core';
-import { MovieForm } from '../../movie.form';
-
-@Component({
-  selector: '[movie] movie-summary-country',
-  templateUrl: './country.component.html',
-  styleUrls: ['./country.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class MovieSummaryCountryComponent implements OnInit {
-  @Input() movie: MovieForm;
-  @Input() link: string;
-
-  constructor(private cdr: ChangeDetectorRef) { }
-
-  ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
-  }
-}
