--- libs/movie/src/lib/movie/form/summary-deprecated/credit/credit.component.ts
@@ -1,24 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input, ChangeDetectorRef, OnInit } from '@angular/core';
-import { CreditForm } from '../../sales-cast/sales-cast.form';
-import { MovieForm } from '../../movie.form';
-
-@Component({
-  selector: '[movie] movie-summary-credit',
-  templateUrl: './credit.component.html',
-  styleUrls: ['./credit.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class MovieSummaryCreditComponent implements OnInit {
-  @Input() movie: MovieForm;
-  @Input() link: string;
-
-  constructor(private cdr: ChangeDetectorRef) {}
-
-  ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
-  }
-
-  public producerHasNoValue(producer: CreditForm) {
-    return !producer.get('firstName').value || !producer.get('lastName').value || !producer.get('role').value;
-  }
-}
