--- libs/movie/src/lib/movie/form/summary-deprecated/evaluation/evaluation.component.ts
@@ -1,19 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input, ChangeDetectorRef, OnInit } from '@angular/core';
-import { MovieForm } from '../../movie.form';
-
-@Component({
-  selector: '[movie] movie-summary-evaluation',
-  templateUrl: './evaluation.component.html',
-  styleUrls: ['./evaluation.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class MovieSummaryEvaluationComponent implements OnInit {
-  @Input() movie: MovieForm;
-  @Input() link: string;
-
-  constructor(private cdr: ChangeDetectorRef) {}
-
-  ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
-  }
-}
