--- libs/movie/src/lib/movie/form/summary-deprecated/sales-cast/sales-cast.component.ts
@@ -1,19 +1,0 @@
-import { Component, ChangeDetectionStrategy, Input, OnInit, ChangeDetectorRef } from '@angular/core';
-import { MovieForm } from '../../movie.form';
-
-@Component({
-  selector: '[movie] movie-summary-sales-cast',
-  templateUrl: './sales-cast.component.html',
-  styleUrls: ['./sales-cast.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class MovieSummarySalesCastComponent implements OnInit {
-  @Input() movie: MovieForm;
-  @Input() link: string;
-
-  constructor(private cdr: ChangeDetectorRef) {}
-
-  ngOnInit() {
-    this.movie.valueChanges.subscribe(_ => this.cdr.markForCheck());
-  }
-}
